Adjustments to UI/ Sound settings/Rhythm game upddate
Made slight adjustments to the position of some UI elements. Sound settings in the main menu actually apply to the game. Made adjustments to the position of note spawners to better align notes to the middle of their respective lanes due to their new size. Slight change to the size of the perfect zone the player has to hit the note. Add reputation as a stat in the rhythm game. Fixed bug where notes weren't falling in the far right line and fixed loading screen hint not changing.