Added Left/Right Hand Mode and Unpause Timer
Player is asked what hand they prefer when starting the game. This changes the Rhythm Game controls to A, S, D and F for left and H, J, K and L for right. Pause button is now set to spacebar and when the player unpauses there is a 3 second timer before the game actually unpauses